Mercs and Missiles
I wanted to show missiles and started by showing the mercs.

Missiles are turned off by default in settings.
There are two types of missiles, Player and Enemy, and both can be turned on/off separately.

As of right now I have 13 types of missiles each with a custom color in settings.

missiles.ini is used to process what missiles are included, ignored, unknown(probably left out of the ini), and 2 per damage type, Major and Minor. Moving the missile identifier in the INI to different subsections will color the missile according the the settings. For example Frozen Orb has 4 missiles;
[IceMajor]
frozenorb=260
frozenorbnova=262
frozenorbexplode=263
...
[IceMinor]
frozenorbbolt=261
...

Or another example Blessed Hammers;
[MagicMajor]
blessedhammer=92
...
[MagicMinor]
blessedhammerex=640
...

There is a debug setting that turn a tooltip on for either "classified" missiles and otherMissiles/unknowns.